+++ app/src/main/java/cn/navitool/theme/BaseThemeController.java
@@ -148,6 +148,7 @@ public abstract class BaseThemeController implements IClusterTheme {
     @Override public void updateTripInfo(float distance, long duration) {}
     @Override public void updateOdometer(float odometer) {}
     @Override public void updateInstantFuel(float fuel) {}
+    @Override public void updateFuelRemain(float fuelLiters) {}
     @Override public void updateIndoorTemp(float temp) {}
     @Override public void updateTrafficLight(TrafficLightInfo info) {}
     @Override public void resetTrafficLights() {}